[Ftp_sync] Fixed.

+++ ftp_sync.php
@@ -20,6 +20,7 @@ define('SYS_NAME', 'ftp_sync'); 				// FTP directroy path
 define('ROOT_PATH', dirname(__FILE__).'/');		// Root Path
 define('RUNTIME_DIR', ROOT_PATH.'runtime/');	// Root Path
 define('FTP_PATH', '/home/iptvftp/'); 			// FTP directroy path
+define('FTP_DONE_PATH', false); 				// FTP done files' directroy path (false: delete files)
 define('CHECK_LIST_PATH', RUNTIME_DIR.SYS_NAME.'_checklist.data');		// File path for save last list data
 define('RUNTIME_PATH', RUNTIME_DIR.SYS_NAME.'_runtime.data');			// File path for script runtime check
 define('FILESIZE_PERIOD_MINUTES', 1);			// Minutes of file size unchanged and exceeded
@@ -86,6 +87,13 @@ $checkList = unserialize( file_get_contents(CHECK_LIST_PATH) );
 
 $newList = [];
 
+
+# AWS S3 Helper
+require dirname(__FILE__).'/libs/AwsS3Helper.class.php';
+$s3Config = require dirname(__FILE__).'/config/s3.php';
+$awsS3Helper = new AwsS3Helper($s3Config);
+
+# Upload batch of checked files
 foreach ($fileList as $key => $file) {
 	
 	$fileName = pathinfo($file, PATHINFO_BASENAME);
@@ -124,11 +132,6 @@ foreach ($fileList as $key => $file) {
 
 				# Check if is the file size unchanged and exceeded FILESIZE_PERIOD_MINUTES
 				if ( (NOW - $lastFile['size_locked_at']) >= (FILESIZE_PERIOD_MINUTES*60) ) {
- 
-					# AWS S3 Helper
-					require dirname(__FILE__).'/libs/AwsS3Helper.class.php';
-					$s3Config = require dirname(__FILE__).'/config/s3.php';
-					$awsS3Helper = new AwsS3Helper($s3Config);
 
 					# Upload configuration
 					$time = date("H:i:s" ,NOW);
@@ -148,9 +151,16 @@ foreach ($fileList as $key => $file) {
 					}
 
 
-					# Remove the local file
+					# Remove or move the local file
 
-					$result = unlink($newFile['path']); 
+					if (FTP_DONE_PATH===false) {
+
+						$result = unlink($newFile['path']); 
+
+					} else {
+
+						rename($newFile['path'], FTP_DONE_PATH.'/'.$fileName);
+					}
 
 					if (!$result) {
